Store application state in localStorage
This stores the current state in the cim_state key, and the session
history (which grows unbounded and is basically never used right now
anyway) in a separate key called cim_session_history.

Because this stuff all needs to be stringified, it's easier to store the
confusion matrix as a JS object rather than a nested mapping.