Recalibrate cat faces
Since 50% is "randomly guessing" (at least in the red/yellow case), it
makes sense to have a higher standard in this case.

For higher levels, we can either keep it this way (on the basis that
they'll already have mastered n-1 of the n chords), or widen the scale (33%
being randomly guessing for the 3-chord case, 25% being randomly
guessing for the 4-chord chase, etc).