Add support for Tone.js
This will change the sound of the chords a little (which is why we still
support the "old" version of the piano for the light mode), but it also
allows us to easily expand colors and instruments.

In the future, it will also allow us to play the chord for an arbitrary
length of time if desired.