Layout improvements
Currently when there are a lot of colors on the screen or when you are
using a phone, there's some chance you might have to do some scrolling,
or there is other wasted space.

This eliminates the big "Chord Identification Method Trainer" text,
moves the controls in-line with the menu items, and hides the menu under
a hamburger button when using small screens.

It also makes some changes to the grid size depending on the screen size
and number of colors.