Add single not trainer related settings
It is not yet clear what the optimal method / UI is going to be for how
to run the single note mode, so I'm erring on the side of giving a lot
of options for how to do it.

The biggest problem with be the transition to the final stage, since
there's no natural "Hey I'm at the next step" level.